~ 68 Props의 심화활용
