+++ React-CH3-37-starting-project/src/components/TabButton/TabButton.jsx
@@ -1,7 +1,7 @@
-export default function TabButton({ children, onSelect, isSelected }) {
+export default function TabButton({ children, isSelected, ...props }) {
     return (
         <li>
-            <button className={isSelected ? "active" : null} onClick={onSelect}>
+            <button className={isSelected ? "active" : null} {...props}>
                 {children}
             </button>
         </li>
