+++ React-CH3-37-starting-project/src/components/Tabs.jsx
@@ -1,0 +1,9 @@
+export default function Tabs({ children, buttons, buttonsContainer = "menu" }) {
+    const ButtonsContainer = buttonsContainer;
+    return (
+        <>
+            <ButtonsContainer>{buttons}</ButtonsContainer>
+            {children}
+        </>
+    );
+}
